poprawki mde nowy certyfikat dla communication
git-svn-id: https://localhost:8443/svn/hm-mde-1@1294 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/src/edrutils/include/utils/Utils.h
@@ -145,7 +145,7 @@ inline static std::string readStream(std::istream & stream)
 	std::string ret;
 
 	std::streamsize read = 0;
-	while ((read = forceReadSome(stream, buffer.get(), BufferSize)) > 0) { ret.append(buffer.get(), read); }
+	while ((read = forceReadSome(stream, buffer.get(), BufferSize)) > 0) { ret.append(buffer.get(), static_cast<unsigned int>(read)); }
 	return std::move(ret);
 }
